Workaround for ts-loader issue with wepack
While no errors get reported in VSCode when opening the project, the package script spits out a lot of TS errors during webpack doing it's work. This workaround works by adding  to the options of ts-loader so that it does not deal with TS verification. The ts-loader configuration also is now shared in all three target configs. This workaround is not optimal, however should be okay as long as code gets linted before packaging.